Update README.md
I added a bit to the FastAPI information to make sure that the application runs ons port 8080. This is a common port and used often. I changed the Flask App code as well to make sure that it will run on localhost port 8000, so that both apps can be run at the same.

Former-commit-id: 9dabce4151490bef3e8ffe043720a44f333b0e7e [formerly 9dabce4151490bef3e8ffe043720a44f333b0e7e [formerly 720f819badde584f8f28983ffb364847eb3b2ff7]]
Former-commit-id: 6c1b134ca70bcd01f97d4f6ad0034132a8f7548b
Former-commit-id: bf2acab153eda18f1c791799e8dfab5d5d49152d