Update README.md
I added a bit to the FastAPI information to make sure that the application runs ons port 8080. This is a common port and used often. I changed the Flask App code as well to make sure that it will run on localhost port 8000, so that both apps can be run at the same.